Fix hashcode for cached box value
Only objects (i.e., non-primitive types) can be hashed. If this object
is not allocated on the heap, it must be a compiler-generated cached
boxing value (when there is no escape analysis). These cached values
reside in the program's text section, which is immutable and non-movable.

Therefore, we can safely return the object's address as its hash value.